working prototype

+++ server/src/api/students/router.py
@@ -12,29 +12,24 @@ from server.src.dao.services import balancer
 students_route = APIRouter(prefix="/students")
 
 
-# TODO сделать выдачу только 10 результатов
 @students_route.get("/", summary="Получить список студентов отфильтрованный по параметрам")
 async def get_students(
         filter_query: Annotated[FilterStudents, Query()],
         session: AsyncSession = Depends(get_async_session)
-):  # TODO сделать схему returnа
+):
     return await StudentDAO.find_all(session, filter_query.model_dump(exclude_none=True))
 
-
-# TODO проверить что returnится
 
 @students_route.get("/{student_id}", summary="Получить одного студента по id", response_model=StudentRead)
 async def get_student_by_id(
         student_id: int,
         session: AsyncSession = Depends(get_async_session)
-):  # TODO -> SInstructorCard
+):
     student = await StudentDAO.find_one_or_none_by_id(session, student_id)
     if not student:
         raise HTTPException(status_code=404, detail="Студент не найден")
     return student
 
-
-# TODO проверить что returnится
 
 @students_route.post("/add", summary="Добавление студента")
 async def add_student(
@@ -48,7 +43,7 @@ async def add_student(
             return {"message": "Нельзя зачислять студента на кафедру без преподавателей"}
         added = await StudentDAO.add(session, **student.model_dump())
         if added:
-            # background_tasks.add_task(balancer.balance, sync_session, added.department_id)
+            background_tasks.add_task(balancer.balance, sync_session, added.department_id)
             return {"message": "Студент успешно добавлен!", "student": student}
         else:
             return {"message": "Ошибка при добавлении студента!"}
@@ -58,6 +53,8 @@ async def add_student(
 async def update_student(
         student_id: int,
         upd_data: SStudentUpd,
+        background_tasks: BackgroundTasks,
+        sync_session=Depends(get_sync_session),
         session: AsyncSession = Depends(get_async_session)
 ):
     async with session.begin():
@@ -78,15 +75,27 @@ async def update_student(
             updated = True
 
         if updated:
+            if upd_data.department_id:
+                background_tasks.add_task(balancer.balance, sync_session, upd_data.department_id)
             return {"message": "Данные студента успешно обновлены!", "student": updated}
         else:
             return {"message": "Ошибка при обновлении данных студента!"}
 
 
 @students_route.delete("/{student_id}/delete")
-async def delete_student(student_id: int, session: AsyncSession = Depends(get_async_session)):
+async def delete_student(
+        student_id: int,
+        background_tasks: BackgroundTasks,
+        sync_session=Depends(get_sync_session),
+        session: AsyncSession = Depends(get_async_session)
+):
     async with session.begin():
+        student = await StudentDAO.find_one_or_none_by_id(session, student_id)
+        if not student:
+            raise HTTPException(status_code=404, detail="Такого студента нет")
+        department_id = student.department_id
         deleted = await StudentDAO.delete_by_id(session, student_id)
         if not deleted:
             raise HTTPException(status_code=404, detail="Ошибка при отчислении")
+        background_tasks.add_task(balancer.balance, sync_session, department_id)
         return {"message": "Студент отчислен"}
